requirements.txt
Django==2.0.2
django-cors-headers==2.1.0
django-rest-framework==0.1.0
djangorestframework==3.7.7
psycopg2==2.7.4
pytz==2018.3
